+++ src/QuizzerApp.Api/Controllers/AuthController.cs
@@ -1,5 +1,6 @@
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using QuizzerApp.Application.Features.Commands.User.CreateUser;
 using QuizzerApp.Application.Features.Commands.User.Login;
 
 namespace QuizzerApp.Api.Controllers;
@@ -15,12 +16,24 @@ public class AuthController : ControllerBase
         _sender = sender;
     }
 
-    [HttpPost]
+    [HttpPost("login")]
     public async Task<IActionResult> Login(UserLoginCommand command)
     {
         var res = await _sender.Send(command);
 
         return Ok(res);
-    } 
+    }
+
+    [HttpPost("register")]
+    public async Task<IActionResult> Register(CreateUserCommand command)
+    {
+        var res = await _sender.Send(command);
+
+        if (!res)
+            throw new Exception();
+
+        return StatusCode(201);
+    }
+
 
 }
