+++ src/QuizzerApp.Api/Controllers/UserController.cs
@@ -1,8 +1,5 @@
-using System.Net;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Identity.Client;
-using QuizzerApp.Application.Features.Commands.User.CreateUser;
 
 namespace QuizzerApp.Api.Controllers;
 
@@ -17,15 +14,11 @@ public class UserController : ControllerBase
         _sender = sender;
     }
 
-    [HttpPost]
-    public async Task<IActionResult> Register(CreateUserCommand command)
+    [HttpGet("me")]
+    public async Task<IActionResult> Me()
     {
-        var res = await _sender.Send(command);
-
-        if (res)
-            return StatusCode(201);
-
-        throw new Exception();
+        return Ok("TODO:");
     }
+
 
 }
