+++ src/QuizzerApp.Api/appsettings.Development.json
@@ -7,5 +7,11 @@
   },
   "SqlServer": {
     "ConnectionString": "Server=localhost,1433;database=QuizzerAppDbDemo;user=sa;password=Root1234!;TrustServerCertificate=true"
+  },
+  "JwtSettings": {
+    "secret": "N3bul44p1S3cr3tK3yF0rN3bul44p1S3cr3tK3yF0rN3bul44p1S3cr3tK3yF0r",
+    "issuer": "QuizzerAppDbDemo",
+    "audience": "http://localhost:5174",
+    "expires": 120
   }
 }
